+++ src/Surfnet/ServiceProviderDashboard/Application/CommandHandler/Entity/SaveSamlEntityCommandHandler.php
@@ -52,7 +52,7 @@ class SaveSamlEntityCommandHandler implements CommandHandler
     {
         // If the entity does not exist yet, create it on the fly
         if (is_null($command->getId())) {
-            $id = (string) Uuid::uuid1();
+            $id = Uuid::uuid1()->toString();
             if (!$this->repository->isUnique($id)) {
                 throw new InvalidArgumentException(
                     'The id that was generated for the entity was not unique, please try again'