+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/DashboardBundle/Security/Voter/ManageEntityAccessGrantedVoter.php
@@ -38,16 +38,21 @@ class ManageEntityAccessGrantedVoter extends Voter
 
     protected function supports($attribute, $subject)
     {
-        $supportedAttribute = $attribute === self::MANAGE_ENTITY_ACCESS;
-        $subjectFieldsPresent = isset($subject['manageId']) && isset($subject['environment']);
+        if ($attribute !== self::MANAGE_ENTITY_ACCESS) {
+            return false;
+        }
 
-        return $supportedAttribute && $subjectFieldsPresent;
+        if (!isset($subject['manageId']) || !isset($subject['environment'])) {
+            return false;
+        }
+
+        return true;
     }
 
     protected function voteOnAttribute($attribute, $subject, TokenInterface $token)
     {
         // Administrator is always allowed to delete an entity
-        if (in_array("ROLE_ADMINISTRATOR", $token->getRoles())) {
+        if ($token->hasRole('ROLE_ADMINISTRATOR')) {
             return true;
         }