+++ src/Surfnet/ServiceProviderDashboard/Infrastructure/Manage/Http/Exception/MalformedResponseException.php
@@ -24,6 +24,6 @@ class MalformedResponseException extends HttpException
 {
     public function __construct($message, Exception $previous = null)
     {
-        return parent::__construct($message, 0, $previous);
+        parent::__construct($message, 0, $previous);
     }
 }