+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.1.6
+**Bugfix**
+  * Escape the wysiwyg translations #281
+  * Fix the statuscode for switch to nonexistent service #280
+  * Add additional secyurity headers to response #278
+  * Fix the entity delete acl #276
+  * Fix the translations acl #277
+ 
 ## 2.1.5
 **Bugfix**
  * Ensure the issues are indexed on Manage id #273
@@ -37,7 +45,7 @@
  * Add translatable contact email placeholder #259
  * Rename 'team name' to 'team identifier' #258
  * Add translatable footer links #253
- * Show motivationless attributes on entity detail #255
+ * Show motivationless attributes on entity detaOil #255
  * Show OIDC items on entity details #256
  * Fix urn validator regex #251
  * Fix the playground url after copy to production #247